updating fields in liability calc response model to match hip calc BE model

+++ app/models/liabilitycalculation/EndOfYearEstimate.scala
@@ -22,7 +22,7 @@ case class EndOfYearEstimate(
                               incomeSource: Option[List[IncomeSource]],
                               totalEstimatedIncome: Option[Int] = None,
                               totalTaxableIncome: Option[Int] = None,
-                              totalAllowancesAndDeductions: Option[BigDecimal] = None,
+                              totalAllowancesAndDeductions: Option[Int] = None,
                               incomeTaxAmount: Option[BigDecimal] = None,
                               nic2: Option[BigDecimal] = None,
                               nic4: Option[BigDecimal] = None,
